2024/6/2 포인트 구현 및 좋아요 싫어요시 포인트 감소 증가 구현
매일 첫 로그인시 포인트 10점 및 좋아요 싫어요 +1 -1 구현 완료 나의 페이지 기본적인 틀 구현 완료 오늘 밤에 할 일 board테이블에 writer를 삭제하고 memberId일 때 nickname 주기 jpa는 자동적으로 매핑을 할 수 있는데 확 실히 매우 불편하다..

+++ src/main/java/com/JYProject/project/model/Member.java
@@ -19,21 +19,16 @@ public class Member{
 
     //("아이디")
     private String loginId;
-
     //("비밀번호")
     private String pw;
     //("이름")
     private String name;
     //("닉네임")
     private String nickname;
-
-
     //("핸드폰번호")
     private String phone;
-
     //("이메일")
     private String email;
-
     //("우편번호")
     private String zipCode;
     //("주소")
@@ -49,7 +44,7 @@ public class Member{
     //("등급")
     private String  Grade;
     //("포인트")
-    private Long point;
+    private int point;
     //("프로필이미지")
     private String profileImg;
